+++ main.go
@@ -92,6 +92,7 @@ type server struct {
 	VINTERVAL  uint32
 	VREPEATERR []string
 	VREPEATCNT uint32
+	VREPLACE   bool
 }
 
 // UssAllow : type for key and slice pairs of a virtual host and user/hash allowable pairs
@@ -129,6 +130,7 @@ type GetTask struct {
 	Interval  uint32   `json:"interval"`
 	Repeaterr []string `json:"repeaterr"`
 	Repeatcnt uint32   `json:"repeatcnt"`
+	Replace   bool     `json:"replace"`
 	Stdcode   int      `json:"stdcode"`
 	Stdout    string   `json:"stdout"`
 	Errcode   int      `json:"errcode"`
@@ -149,6 +151,7 @@ type PostTask struct {
 	Interval  uint32   `json:"interval"`
 	Repeaterr []string `json:"repeaterr"`
 	Repeatcnt uint32   `json:"repeatcnt"`
+	Replace   bool     `json:"replace"`
 }
 
 // DelTask : type task for delete video tasks
@@ -165,6 +168,7 @@ type DelTask struct {
 	Interval  uint32   `json:"interval"`
 	Repeaterr []string `json:"repeaterr"`
 	Repeatcnt uint32   `json:"repeatcnt"`
+	Replace   bool     `json:"replace"`
 	Stdcode   int      `json:"stdcode"`
 	Stdout    string   `json:"stdout"`
 	Errcode   int      `json:"errcode"`
@@ -187,6 +191,7 @@ type RawTask struct {
 	Interval  uint32
 	Repeaterr []string
 	Repeatcnt uint32
+	Replace   bool
 	Stdcode   int
 	Stdout    string
 	Errcode   int
@@ -208,6 +213,7 @@ type FullTask struct {
 	Interval  uint32
 	Repeaterr []string
 	Repeatcnt uint32
+	Replace   bool
 	Stdcode   int
 	Stdout    string
 	Errcode   int
@@ -409,6 +415,7 @@ func init() {
 	rgxsslcrt := regexp.MustCompile("^(/?[^/\x00]*)+/?$")
 	rgxsslkey := regexp.MustCompile("^(/?[^/\x00]*)+/?$")
 	rgxshell := regexp.MustCompile("^(/?[^/\x00]*)+/?$")
+	rgxreplace := regexp.MustCompile("^(?i)(true|false)$")
 
 	for _, Server := range config.Server {
 
@@ -573,6 +580,9 @@ func init() {
 		mchvrepeatcnt := RBUint(Server.VREPEATCNT, 0, 1000)
 		Check(mchvrepeatcnt, section, "vrepeatcnt", fmt.Sprintf("%d", Server.VREPEATCNT), "from 0 to 1000", DoExit)
 
+		mchvreplace := rgxreplace.MatchString(fmt.Sprintf("%t", Server.VREPLACE))
+		Check(mchvreplace, section, "vreplace", fmt.Sprintf("%t", Server.VREPLACE), "true or false", DoExit)
+
 		// Output Important Server Configuration Options
 
 		appLogger.Warnf("| Host [%s] | Shell [%s]", Server.HOST, Server.SHELL)
@@ -582,6 +592,13 @@ func init() {
 		appLogger.Warnf("| Host [%s] | Scheduler Task TTL Seconds [%d]", Server.HOST, Server.VTTLTIME)
 		appLogger.Warnf("| Host [%s] | Scheduler Task Interval Seconds [%d]", Server.HOST, Server.VINTERVAL)
 		appLogger.Warnf("| Host [%s] | Scheduler Task Repeat Tries [%d]", Server.HOST, Server.VREPEATCNT)
+
+		switch {
+		case Server.VREPLACE:
+			appLogger.Warnf("| Host [%s] | Scheduler Task Replace [ENABLED]", Server.HOST)
+		default:
+			appLogger.Warnf("| Host [%s] | Scheduler Task Replace [DISABLED]", Server.HOST)
+		}
 
 	}