+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Version: 1.1.3
+========
+
+Added in version 1.1.3:
+
+- Field replace for /task
+- Parameter ```vreplace```
+
 Version: 1.1.2
 ========